Crate lithium

lithium

lithium is (probably) fast and (hopefully) ergonomic SQL builder.
Attention: stuff is not even close to ready and can break in a flash. Also, everything is built on nightly and wasn't checked on stable/beta.

As for now, lithium provides interface to create a SQL, but it's not responsible for executing it.
lithium does not handle security in any way, it's on your DB driver!

You can find examples in documentation for every struct.

Modules

common
Keeps stuff (mostly traits) that is used (or is going to be) across different queries.

insert
Keeps INSERT related stuff.

select
Keeps SELECT related stuff.

update
Keeps UPDATE related stuff.

where_cl
Keeps WHERE related stuff.

Structs

Insert
Represents INSERT query.

Select
Represents SELECT query.

Update
Represents UPDATE query

Where
Represents WHERE clause which is widely used in different queries.

Traits

AsStr
Is used to build up methods which can receive either &str or &Subquery in a convenient way. You can find examples in some of Select's methods.

Pusheable
Is used to build up methods which can receive either &str or &[&str; N] in a convenient way. You can find examples in some of Select's methods.

ToSQL