Module lithium::common

Keeps stuff (mostly traits) that is used (or is going to be) across different queries.

Structs

Subquery
Struct that is used to keep result from to_sql of some query. If you use with_alias - keep in mind that it's changing content of query in irreversible way.
We do this because we need &str to have a nice way of using subqueries and avoid forcing users to use String when they don't really need to.

Traits

AsStr
Is used to build up methods which can receive either &str or &Subquery in a convenient way. You can find examples in some of Select's methods.

Pusheable
Is used to build up methods which can receive either &str or &[&str; N] in a convenient way. You can find examples in some of Select's methods.

ToSQL