Struct lithium::common::Subquery

pub struct Subquery<'a> {
    pub query: String,
    // some fields omitted
}

Struct that is used to keep result from to_sql of some query. If you use with_alias - keep in mind that it's changing content of query in irreversible way.
We do this because we need &str to have a nice way of using subqueries and avoid forcing users to use String when they don't really need to.

Fields

query

Keeps generated SQL

Methods

impl<'a> Subquery<'a>

fn new(query: String) -> Self

fn with_alias(self, alias: &'a str) -> Self

Appends AS {alias} to the query field